feat: R2DBC SPI upgrade to 1.0.0.RELEASE (#726)
* Updated r2dbc.version to 1.0.0.RELEASE

* Code changes for r2dbc version upgrade to 1.0.0.RELEASE

* fix getRowsUpdatedTest

* fixed ITs

* removed module and updated README.md

* removed cloud-spanner-spring-data-r2dbc

* removed cloud-spanner-spring-data-r2dbc-sample

+++ cloud-spanner-r2dbc/src/test/java/com/google/cloud/spanner/r2dbc/it/SpannerQueryUtil.java
@@ -43,10 +43,10 @@ class SpannerQueryUtil {
   /**
    * Executes a DML query and returns the rows updated.
    */
-  static int executeDmlQuery(Connection connection, String sql) {
+  static long executeDmlQuery(Connection connection, String sql) {
 
     Mono.from(connection.beginTransaction()).block();
-    int rowsUpdated = Mono.from(connection.createStatement(sql).execute())
+    long rowsUpdated = Mono.from(connection.createStatement(sql).execute())
         .flatMap(result -> Mono.from(result.getRowsUpdated()))
         .block();
     Mono.from(connection.commitTransaction()).block();
